ips/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
Babel==2.8.0
bcrypt==3.2.2
boto3==1.35.35
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==2.8
ipaddr==2.2.0
Jinja2==3.0.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
netaddr==0.8.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2022.1
PyYAML==6.0.1
redis==5.1.0
requests==2.22.0
rich==13.9.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==1.25.11
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
PyJWT==2.9.0
